requirements.txt
Pillow==7.2.0 
colorama==0.4.3  
